Add files via upload
This project is a web-based application that allows you to generate QR codes from any text input. It is built with HTML, CSS, and JavaScript and uses a simple and user-friendly interface. You can scan the QR code with your mobile or any QR code scanner and display the information. You can view the source code and a live demo of the project from this repository.